pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "sam2-contrails"
version = "0.1.0"
description = "SAM2-based contrail detection, tracking, and attribution from all-sky camera imagery"
readme = "README.md"
license = { file = "LICENSE" }
authors = [{ name = "Ramon Dalmau" }]
keywords = [
    "contrail",
    "segmentation",
    "tracking",
    "attribution",
    "SAM2",
    "aviation",
    "atmospheric-science",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: European Union Public Licence 1.2 (EUPL 1.2)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Atmospheric Science",
]
requires-python = ">=3.10"
dependencies = [
    "torch>=2.5.1",
    "torchvision>=0.20.1",
    "numpy>=1.24",
    "Pillow>=9.4",
    "tqdm>=4.66",
    "hydra-core>=1.3.2",
    "iopath>=0.1.10",
    # CLI (contrailtrack run / evaluate / generate-prompts)
    "typer>=0.9",
    "structlog>=24.0",
]

[project.urls]
Homepage       = "https://github.com/ramondalmau/sam2-contrails"
Repository     = "https://github.com/ramondalmau/sam2-contrails"
"Bug Tracker"  = "https://github.com/ramondalmau/sam2-contrails/issues"
"Hugging Face" = "https://huggingface.co/ramondalmau/sam2-contrails"

[project.scripts]
contrailtrack = "contrailtrack.cli:app"

[project.optional-dependencies]
inference = [
    "pycocotools>=2.0.8",
    "opencv-python>=4.7",
    "pandas>=2.0,<3",
    "pyarrow>=14.0",
    "scipy>=1.10,<1.16",
    "huggingface_hub>=0.20",
]
prompts = [
    "pycontrails[ecmwf]>=0.51,<0.60",
    "geopandas>=0.14",
    "rasterio>=1.3",
    "shapely>=2.0",
    "xarray>=2023.1",
    "netcdf4>=1.6",
    "geographiclib>=2.0",
]
all = ["sam2-contrails[inference,prompts]"]
dev = [
    "pytest>=7.0",
    "black>=24.0",
]

[tool.setuptools.packages.find]
include = ["sam2*", "contrailtrack*", "training*"]

[tool.setuptools.package-data]
"sam2"          = ["configs/**/*.yaml"]
"contrailtrack" = ["configs/**/*.yaml"]

# ── PyTorch CUDA index (uv-native configuration) ──────────────────────────────
# Pre-configured for CUDA 12.8. If your system uses a different version,
# change "cu128" below (e.g. "cu121" for CUDA 12.1) before running uv sync.
# For CPU-only, remove this entire block.

[[tool.uv.index]]
name = "pytorch-cu128"
url  = "https://download.pytorch.org/whl/cu128"
explicit = true

[tool.uv.sources]
torch       = [{ index = "pytorch-cu128", marker = "sys_platform != 'darwin'" }]
torchvision = [{ index = "pytorch-cu128", marker = "sys_platform != 'darwin'" }]
